Implement shopping list feature with Spoonacular integration and DTO mapping
- Created ShoppingItem entity and repository
- Implemented ShoppingListService with auto-generation from MealPlan
- Fetched real ingredients via Spoonacular recipe details
- Added endpoints for listing, adding, removing, and checking items
- Introduced ShoppingItemRequest and ShoppingItemResponse DTOs
- Applied mapping to clean up response payloads

+++ backend/recipedia/src/main/java/com/matildaerenius/controller/RecipeController.java
@@ -50,7 +50,7 @@ public class RecipeController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<?> getRecipeDetails(@PathVariable int id) {
+    public ResponseEntity<?> getRecipeDetails(@PathVariable Long id) {
         try {
             String detailsJson = recipeService.getRecipeDetails(id);
             return ResponseEntity.ok(objectMapper.readTree(detailsJson));
